refactor: 리프레시 토큰 만료 기한 타입 캐스팅

+++ src/main/java/com/playkuround/playkuroundserver/domain/auth/token/application/TokenManager.java
@@ -51,7 +51,7 @@ public class TokenManager {
     }
 
     public Date createRefreshTokenExpirationTime() {
-        return new Date(System.currentTimeMillis() + Long.parseLong(refreshTokenExpirationTime));
+        return new Date(System.currentTimeMillis() + (long) Double.parseDouble(refreshTokenExpirationTime));
     }
 
     public String createAccessToken(String email, Date expiredAt) {
